pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]

[project]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering"
]
dependencies = [
  "dash>=2.0.0",
  "dash-bootstrap-components>=1.0.0",
  "dash-cytoscape>=0.3.0",
  "cantera>=3.0.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0",
  "pydantic>=2.0.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "python-multipart>=0.0.6"
]
description = "A visual interface for Cantera reactor networks"
dynamic = ["version"]
license = {file = "LICENSE"}
name = "boulder"
readme = "README.md"

[project.optional-dependencies]
dev = [
  "mypy>=1.0.0",
  "types-PyYAML>=6.0.0",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.23.0"
]

[project.scripts]
boulder = "boulder.cli:main"
sim2stone = "boulder.sim2stone_cli:main"
stone2sim = "boulder.stone2sim_cli:main"

[tool.coverage.run]
branch = true
concurrency = ["multiprocessing"]
parallel = true

[tool.mypy]
strict = true

[tool.pytest.ini_options]
# Add useful options
addopts = [
  "--strict-markers",
  "--strict-config",
  "--verbose",
  "--tb=short",
  "--durations=10"
]
# Filter warnings
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning"
]
# Test markers
markers = [
  "e2e: marks tests as end-to-end tests requiring Playwright (deselect with '-m \"not e2e\"')",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
  "unit: marks tests as unit tests"
]
# Minimum version requirements
minversion = "6.0"
python_classes = ["Test*"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
# Test discovery patterns
testpaths = ["tests"]

[tool.ruff]
indent-width = 4
line-length = 88

[tool.ruff.lint]
ignore = ["D1"]
select = ["F", "E", "W", "I", "D"]

[tool.ruff.lint.pycodestyle]
max-line-length = 110

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.setuptools]
packages = ["boulder"]

[tool.setuptools_scm]
write_to = "boulder/version.py"
write_to_template = '''
# Do not change! Do not track in version control!
__version__ = "{version}"
'''
